Bunch More Work: RAW improvements, sine wave generator.
Loading raw PCM is now hooked up to the public API, and offers both IOStream
loading and creation directly from a memory buffer (with or without free'ing
the buffer on destruction, so you can fully replace the zero-copy magic of
SDL2_mixer's Mix_QuickLoad_RAW).

The sine wave generator is useful if you just need _something_ to play. It
functions as a dirt-simple audio decoder that doesn't load from an IOStream,
but since it backs a Mix_Audio, it can be assigned to a Mix_Track to
generate procedural sound on demand without the overhead of an SDL_AudioStream.

Just sort of fun to have it in the toolbox.