mpg123: Remove test code to search for an MPEG audio frame header.
In practice, this _did_ find a few valid MP3s in my collection that had some
garbage upfront--some podcasts from 2006, things that probably came from
Napster in 1999 or something, but all of those also had ID3 tags anyhow,
making it unnecessary.

But it _also_ thought sample.ogg was an MP3.

So for now, it's either "it has ID3 tags" or "the first 4 bytes are likely an
MPEG frame header" as the heuristic.