mixer: Redesigned core mixer to support mix groups.
This moved from one bound audio stream per track to one bound audio stream,
total, moving the actual mixing into SDL_remixer and not having SDL3 do it.

This was necessitated by the addition of MIX_Group, which allows a callback
when an abitrary set of tracks have been mixed into a single buffer (to then
be mixed into the final output). This wasn't reasonable to do entirely in SDL,
since we can't guarantee the order of audio stream mixing, or group them,
outside of abusing the logical device mechanism or binding a stream per-group.
(But if you're going to bind per-group, you still have to mix tracks yourself,
so you might as well take complete control.)

This also simplified the locking mechanisms significantly.

Of course, this is a lot of code changes, so hopefully this isn't wildly
unstable now.

While I was in here, I also changed the track "mix" callback into a "raw" and
"cooked" callback: the first for the unmodified data as it is decoded, the
other for the track's buffer after all transformations (gain, frequency, fade,
etc), right before it is mixed.