api: Added MIX_SetTrackIOStream() and a few other things.
This touched just about _everything_, but in many cases it caused code to
simplify. It also exposed some bugs that got fixed, some optimizations, and
a few minor feature addtions (like the ability to skip metadata tag searching
if you know you won't have any on your file).

The WAV, AIFF, and VOC decoders were a casualty of this work; they will return.

Likely there are some new and exciting bugs to be shaken out, too, but this
is progress, and progress is sometimes painful.  :)