mixer: Mark non-device audiostreams as !auto_cleanup.
We are sure to manually clean them up in MIX_DestroyMixer, etc.

This is so someone can use SDL_remixer on a platform where they might call
SDL_Quit() and reinitialize the subsystem with a different audio backend; in
this case, MIX_CreateMixer()-generated mixers should continue to function,
since there's nothing backend-specific about it.

This requires a new property in SDL3's revision control.

For now, we're `#define`ing the property, since this is special-case
functionality, but once SDL 3.4.0 is finalized, we'll remove that define.